chore: swap react spring with framer motion
